Using a map instead of an array of articles.
Using a map will simplify the process of getting an article because I can access an article by using a key instead of a loop.
Fixing issue in the "createArticle" function. Using "r.Header.Values('Content-Type')" instead of "r.Header.Get('application/json')" because the second one caused me errors even if the body of my HTTP request was in the JSON format required.